move command to logic namespace
