+++ src/logic/commands/activate-command.js
@@ -1,5 +1,5 @@
 import { Command } from "./command.js";
-import { CommandType } from '../enums/command-type.js';
+import { CommandType } from '/src/models/game-elements/enums/command-type.js';
 
 export class ActivateCommand extends Command {
     constructor(app, gameState, context) {
