+++ src/logic/commands/move-command.js
@@ -1,4 +1,4 @@
-import { Command } from "./command";
+import { Command } from "./command.js";
 
 export class MoveCommand extends Command {
