+++ src/logic/commands/select-command.js
@@ -1,4 +1,4 @@
-import { Command } from "./command";
+import { Command } from "./command.js";
 
 export class SelectCommand extends Command {
